Update UserDaoImpl.java
Hey there, me again. It looks like you have a lot of comments that look like a dairy. Most importantly, in the method getUserAdress you have a comment saying that it only works if a condition is meet, This is a moment to use "Introduce Assertion"
Sorry if this comes off as mean, this is part of a homework, so bear with me. :)